requirement.txt
django
pydub
pyaudio
scipy
tqdm
tensorflow
pillow
matplotlib
IPython
ipywidgets
